backend/requirements.txt
# FastAPI 核心依赖
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# 数据库
sqlalchemy==2.0.25
alembic==1.13.1
pymysql==1.1.0
cryptography==42.0.0

# Pydantic 设置管理
pydantic==2.5.3
pydantic-settings==2.1.0

# WebSocket 支持
websockets==12.0

# CORS
python-dotenv==1.0.0

# 用户认证
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# 可选：Redis（用于缓存和 WebSocket）
# redis==5.0.1

# 可选：PostgreSQL 驱动
# psycopg2-binary==2.9.9

# 开发工具
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
